+++ src/main/java/com/github/dockerunit/discovery/consul/annotation/impl/TCPHealthCheckExtensionInterpreter.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 import com.github.dockerjava.api.command.CreateContainerCmd;
 import com.github.dockerunit.annotation.ExtensionInterpreter;
-import com.github.dockerunit.discovery.consul.annotation.EnableConsul;
+import com.github.dockerunit.discovery.consul.annotation.TCPHealthCheck;
 import com.github.dockerunit.internal.ServiceDescriptor;
 
-public class EnableConsulExtensionInterpreter implements ExtensionInterpreter<EnableConsul> {
+public class TCPHealthCheckExtensionInterpreter implements ExtensionInterpreter<TCPHealthCheck> {
 
 	private static final String SERVICE_NAME_SUFFIX = "_NAME";
 	private static final String SERVICE_PREFIX = "SERVICE_";
@@ -19,11 +19,11 @@ public class EnableConsulExtensionInterpreter implements ExtensionInterpreter<En
 
 
 	@Override
-	public CreateContainerCmd build(ServiceDescriptor td, CreateContainerCmd cmd, EnableConsul ec) {
-		final String serviceNameEnv = SERVICE_PREFIX + ec.exposedPort() + SERVICE_NAME_SUFFIX + "=" + td.getNamed().value();
+	public CreateContainerCmd build(ServiceDescriptor td, CreateContainerCmd cmd, TCPHealthCheck ec) {
+		final String serviceNameEnv = SERVICE_PREFIX + ec.port() + SERVICE_NAME_SUFFIX + "=" + td.getNamed().value();
 		final String serviceCheckIntervalEnv = SERVICE_CHECK_INTERVAL + "=" + ec.pollingInterval() + "s";
 		final String serviceInitialStatusEnv = SERVICE_CHECK_INITIAL_STATUS + "=" + ec.checkInitialStatus().toString().toLowerCase();
-		final String serviceCheckTCP = SERVICE_CHECK_TCP + "=$SERVICE_IP" + ":" + ec.exposedPort();
+		final String serviceCheckTCP = SERVICE_CHECK_TCP + "=$SERVICE_IP" + ":" + ec.port();
 		
 		List<String> finalEnv = new ArrayList<>();
 		List<String> enableConsulEnv = Arrays.asList(serviceNameEnv, serviceCheckIntervalEnv, serviceInitialStatusEnv, serviceCheckTCP);