+++ src/main/java/com/github/dockerunit/discovery/consul/annotation/impl/UseConsulDnsExtensionInterpreter.java
@@ -1,13 +1,10 @@
 package com.github.dockerunit.discovery.consul.annotation.impl;
 
-import static com.github.dockerunit.discovery.consul.ConsulDiscoveryConfig.DOCKER_BRIDGE_IP_DEFAULT;
-import static com.github.dockerunit.discovery.consul.ConsulDiscoveryConfig.DOCKER_BRIDGE_IP_PROPERTY;
-
 import com.github.dockerjava.api.command.CreateContainerCmd;
 import com.github.dockerjava.api.model.HostConfig;
+import com.github.dockerunit.annotation.ExtensionInterpreter;
 import com.github.dockerunit.discovery.consul.ContainerUtils;
 import com.github.dockerunit.discovery.consul.annotation.UseConsulDns;
-import com.github.dockerunit.annotation.ExtensionInterpreter;
 import com.github.dockerunit.internal.ServiceDescriptor;
 
 import java.util.ArrayList;
@@ -26,7 +23,7 @@ public class UseConsulDnsExtensionInterpreter implements ExtensionInterpreter<Us
 
         List<String> dnsList = new ArrayList<>();
         List<String> currentDnsList = Arrays.asList(Optional.ofNullable(cmd.getHostConfig().getDns()).orElse(new String[0]));
-        String consulDns = dnsIp.orElse(System.getProperty(DOCKER_BRIDGE_IP_PROPERTY, DOCKER_BRIDGE_IP_DEFAULT));
+        String consulDns = dnsIp.orElseThrow(() -> new RuntimeException("Could not detect Consul container ip. Please make sure Consul is running."));
         dnsList.addAll(currentDnsList);
         dnsList.add(consulDns);